airflow_data/include/requirements.txt
apache-airflow==2.6.3
beautifulsoup4==4.10.0
docker
lxml
numpy==1.21.4
pandas==1.3.4
psycopg2-binary==2.9.1
pydantic==1.10.8
python-dotenv==1.0.0
python_dateutil==2.8.1
Requests==2.31.0
selenium==4.9.1